API Reference

SPHtoGrid.Cubic

SPHtoGrid.Quintic

SPHtoGrid.SPHKernel

SPHtoGrid.Smac1ImageInfo

SPHtoGrid.WendlandC4

SPHtoGrid.WendlandC6

SPHtoGrid.mappingParameters

SPHtoGrid.check_center_and_move_particles

SPHtoGrid.check_in_image

SPHtoGrid.domain_decomposition

SPHtoGrid.filter_particles_in_image

SPHtoGrid.find_max_pixel

SPHtoGrid.find_min_pixel

SPHtoGrid.get_d_hsml_2D

SPHtoGrid.get_d_hsml_3D

SPHtoGrid.kernel_value_2D

SPHtoGrid.kernel_value_2D

SPHtoGrid.kernel_value_2D

SPHtoGrid.kernel_value_2D

SPHtoGrid.kernel_value_3D

SPHtoGrid.kernel_value_3D

SPHtoGrid.kernel_value_3D

SPHtoGrid.kernel_value_3D

SPHtoGrid.read_smac1_binary_image

SPHtoGrid.read_smac1_binary_info

SPHtoGrid.sphMapping

SPHtoGrid.sphMapping_2D

SPHtoGrid.sphMapping_3D

SPHtoGrid.write_smac1_par

SPHtoGrid.write_smac2_par

Exported Functions

SPHtoGrid.filter_particles_in_image — Method

filter_particles_in_image(x, hsml, param::mappingParameters)

Checks if a particle is contained in the image and returns an array of Bool.

SPHtoGrid.kernel_value_2D — Method

kernel_value_2D(kernel::Cubic, u::Float64, h_inv::Float64)

Evaluate cubic spline at position $u = rac{x}{h}$.

SPHtoGrid.kernel_value_2D — Method

kernel_value_2D(kernel::Quintic, u::Float64, h_inv::Float64)

Evaluate quintic spline at position $u = rac{x}{h}$.

SPHtoGrid.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC4, u::Float64, h_inv::Float64)

Evaluate WendlandC4 spline at position $u = rac{x}{h}$.

SPHtoGrid.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC6, u::Float64, h_inv::Float64)

Evaluate WendlandC6 spline at position $u = rac{x}{h}$.

SPHtoGrid.kernel_value_3D — Method

kernel_value_3D(kernel::Cubic, u::Float64, h_inv::Float64)

Evaluate cubic spline at position $u = rac{x}{h}$.

SPHtoGrid.kernel_value_3D — Method

kernel_value_3D(kernel::Quintic, u::Float64, h_inv::Float64)

Evaluate quintic spline at position $u = rac{x}{h}$.

SPHtoGrid.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC4, u::Float64, h_inv::Float64)

Evaluate WendlandC4 spline at position $u = rac{x}{h}$.

SPHtoGrid.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC6, u::Float64, h_inv::Float64)

Evaluate WendlandC6 spline at position $u = rac{x}{h}$.

SPHtoGrid.read_smac1_binary_image — Method

read_smac1_binary_image(filename::String)

Reads a binary image file from Smac and returns a Matrix with the pixel values.

SPHtoGrid.read_smac1_binary_info — Method

read_smac1_binary_info(filename::String)

Returns the image info in a Smac1ImageInfo struct.

SPHtoGrid.sphMapping — Method

sphMapping(Pos, HSML, M, ρ, Bin_Quant;
        param::mappingParameters,
        kernel::SPHKernel[,
        show_progress::Bool=true,
        conserve_quantities::Bool=true,
        parallel::Bool=true,
        filter_particles::Bool=true,
        dimensions::Int=2])

Maps the data in Bin_Quant to a grid. Parameters of mapping are supplied in param and the kernel to be used in kernel.

Arguments

Pos: Array with particle positions.

HSML: Array with particle hsml.

M: Array with particle masses.

ρ: Array with particle densities.

Bin_Quant: Array with particle quantity to be mapped.

kernel::SPHKernel: Kernel object to be used.

show_progress::Bool=true: Show progress bar.

conserve_quantities::Bool=true: If quantities should be conserved while mapping, like in Smac (Dolag et. al. 2005).

parallel::Bool=true: Run on multiple processors.

filter_particles::Bool=true: Find the particles that are actually contained in the image.

dimensions::Int=2: Number of mapping dimensions (2 = to grid, 3 = to cube).

SPHtoGrid.write_smac1_par — Function

write_smac1_par([...])

Writes a Smac parameter file. Not all relevant parameters implemented yet!

SPHtoGrid.write_smac2_par — Function

write_smac2_par([...])

Writes a P-Smac2 parameter file. Not all relevant parameters implemented yet!

Exported Types

SPHtoGrid.Cubic — Type

struct Cubic <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
end

Datatype for cubic sph spline.

SPHtoGrid.Quintic — Type

struct Quintic <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
end

Datatype for quintic sph spline.

SPHtoGrid.WendlandC4 — Type

struct WendlandC4 <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
end

SPHtoGrid.WendlandC6 — Type

struct WendlandC6 <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
end

SPHtoGrid.mappingParameters — Type

struct mappingParameters
    x_lim::Vector{Float64}
    y_lim::Vector{Float64}
    z_lim::Vector{Float64}
    center::Vector{Float64}
    pixelSideLength::Float64
    pixelArea::Float64
    Npixels::Vector{Int64}
    x_size::Float64
    y_size::Float64
    z_size::Float64
end

Constructor:

mappingParameters(;x_lim::Vector{Float64}   = [-1.0, -1.0],
                   y_lim::Vector{Float64}   = [-1.0, -1.0],
                   z_lim::Vector{Float64}   = [-1.0, -1.0],
                   center::Vector{Float64}  = [-1.0, -1.0, -1.0],
                   x_size::Float64          =  -1.0,
                   y_size::Float64          =  -1.0,
                   z_size::Float64          =  -1.0,
                   pixelSideLength::Float64 =  -1.0,
                   Npixels::Int64           =   0)

Parameter object for sph to grid mapping. Define either *_lim, or center and *_size. Resolution is defined by pixelSideLength or Npixels.

Private Functions

SPHtoGrid.check_center_and_move_particles — Method

check_center_and_move_particles(x, par::mappingParameters)

Mapping only works if all coordinates are positive. This function shifts the particles into a positive coordinate region.

SPHtoGrid.check_in_image — Method

check_in_image( pos::Float64, hsml::Float64,
                minCoords::Float64, maxCoords::Float64 )

Checks if a particle is in the image frame.

SPHtoGrid.domain_decomposition — Method

domain_decomposition(N::Int64, N_workers::Int64)

Calculate relevant array slices for each worker. Could be done better!

SPHtoGrid.find_max_pixel — Method

find_max_pixel( pos::Float64, hsml::Float64,
                minCoords::AbstractFloat, pixsize_inv::Float64, 
                max_pixel::Integer )

Computes the maximum pixel to which the particle contributes.

SPHtoGrid.find_min_pixel — Method

find_min_pixel( pos::Float64, hsml::Float64,
                minCoords::Float64,
                pixsize_inv::Float64 )

Computes the minimum pixel to which the particle contributes.

SPHtoGrid.get_d_hsml_2D — Method

get_d_hsml_2D( dx::Float64, dy::Float64,
               hsml_inv::Float64 )

Computes the distance in 2D to the pixel center in units of the kernel support.

SPHtoGrid.get_d_hsml_3D — Method

get_d_hsml_3D( dx::Float64, dy::Float64, dz::Float64,
               hsml_inv::Float64 )

Computes the distance in 3D to the pixel center in units of the kernel support.

SPHtoGrid.sphMapping_2D — Method

sphMapping2D(Pos, HSML, M, ρ, BinQuant; param::mappingParameters, kernel, conservequantities::Bool=false, showprogress::Bool=false)

Underlying function to map SPH data to a 2D grid.

SPHtoGrid.sphMapping_3D — Method

sphMapping3D(Pos, HSML, M, ρ, BinQuant; param::mappingParameters, kernel, conservequantities::Bool=false, showprogress::Bool=false)

Underlying function to map SPH data to a 3D grid.

Private Types

SPHtoGrid.SPHKernel — Type

SPHKernel

Supertype for all SPH kernels.

SPHtoGrid.Smac1ImageInfo — Type

struct Smac1ImageInfo

Stores the information in a Smac binary image header.