API Reference

SPHtoGrid.Smac1ImageInfo

SPHtoGrid.mappingParameters

SPHtoGrid.CS14_acc

SPHtoGrid.EpsNtherm

SPHtoGrid.KR07_acc

SPHtoGrid.KR13_acc

SPHtoGrid.P16_acc

SPHtoGrid.Ryu19_acc

SPHtoGrid.Tcmb

SPHtoGrid.analytic_synchrotron_emission

SPHtoGrid.calculate_index_2D

SPHtoGrid.calculate_index_3D

SPHtoGrid.calculate_weights_2D

SPHtoGrid.calculate_weights_3D

SPHtoGrid.check_center_and_move_particles

SPHtoGrid.check_in_image

SPHtoGrid.comptonY

SPHtoGrid.cre_spec_norm_particle

SPHtoGrid.density_2D

SPHtoGrid.domain_decomposition

SPHtoGrid.dsa_spectral_index

SPHtoGrid.euler_matrix

SPHtoGrid.filter_particles_in_image

SPHtoGrid.find_position_periodic

SPHtoGrid.get_d_hsml_2D

SPHtoGrid.get_d_hsml_3D

SPHtoGrid.get_dxyz

SPHtoGrid.get_ijk_min_max

SPHtoGrid.get_map_grid_2D

SPHtoGrid.get_map_grid_3D

SPHtoGrid.get_xyz

SPHtoGrid.kSzPrefac

SPHtoGrid.kinetic_SZ

SPHtoGrid.kr_fitting_function

SPHtoGrid.part_weight_XrayBand

SPHtoGrid.part_weight_emission

SPHtoGrid.part_weight_one

SPHtoGrid.part_weight_physical

SPHtoGrid.part_weight_spectroscopic

SPHtoGrid.project_along_axis

SPHtoGrid.project_along_axis!

SPHtoGrid.project_along_axis!

SPHtoGrid.read_fits_image

SPHtoGrid.read_smac1_binary_image

SPHtoGrid.read_smac1_binary_info

SPHtoGrid.reduce_image_2D

SPHtoGrid.reduce_image_3D

SPHtoGrid.rotate_3D

SPHtoGrid.rotate_3D!

SPHtoGrid.rotate_3D_quantity

SPHtoGrid.rotate_to_xz_plane!

SPHtoGrid.rotate_to_xz_plane!

SPHtoGrid.rotate_to_yz_plane!

SPHtoGrid.rotate_to_yz_plane!

SPHtoGrid.sphMapping

SPHtoGrid.sphMapping

SPHtoGrid.sphMapping_2D

SPHtoGrid.sphMapping_3D

SPHtoGrid.tSzPrefac

SPHtoGrid.thermal_SZ

SPHtoGrid.update_image

SPHtoGrid.write_fits_image

SPHtoGrid.write_smac1_par

SPHtoGrid.write_smac2_par

SPHtoGrid.x_ray_emission

Exported Functions

SPHtoGrid.analytic_synchrotron_emission — Method

analytic_synchrotron_emission( rho_cgs::Array{<:Real}, B_cgs::Array{<:Real},
                               T_K::Array{<:Real}, Mach::Array{<:Real};
                               xH::Real=0.76, dsa_model::Integer=1, ν0::Real=1.4e9,
                               integrate_pitch_angle::Bool=true )

Computes the analytic synchrotron emission with the simplified approach described in Longair Eq. 8.128.

Arguments

rho_cgs::Array{<:Real}: Density in $g/cm^3$.

B_cgs::Array{<:Real}: Magnetic field in Gauss.

T_K::Array{<:Real}: Temperature in Kelvin.

Mach::Array{<:Real}: Mach number.

Keyword Arguments

xH::Float64 = 0.76: Hydrogen fraction of the simulation, if run without chemical model.

dsa_model::Integer=1: Diffuse-Shock-Acceleration model. Takes values 0...4, see next section.

ν0::Real=1.4e9: Observation frequency in $Hz$.

integrate_pitch_angle::Bool=true: Integrates over the pitch angle as in Longair Eq. 8.87.

DSA models

0: KR07_acc. Efficiency model from Kang, Ryu, Cen, Ostriker 2007, http://arxiv.org/abs/0704.1521v1.

1: KR13_acc. Efficiency model from Kang&Ryu 2013, doi:10.1088/0004-637X/764/1/95 .

2: Ryu19_acc. Efficiency model from Ryu et al. 2019, https://arxiv.org/abs/1905.04476 .

3: CS14_acc. Efficiency model from Caprioli&Spitkovsky 2015, doi: 10.1088/0004-637x/783/2/91 .

4: P16_acc. Constant efficiency as in Pfrommer+ 2016, doi: 10.1093/mnras/stw2941 .

SPHtoGrid.density_2D — Function

density_2D( rho::Real, pixelSideLength::Real, 
            Mass::Real=1.989e43, Length::Real=3.085678e21)

Computes the 2D density.

Returns

2D surface density in untis of [g/cm^2].

SPHtoGrid.filter_particles_in_image — Method

filter_particles_in_image(x, hsml, param::mappingParameters)

Checks if a particle is contained in the image and returns an array of Bool.

SPHtoGrid.get_map_grid_2D — Method

get_map_grid_2D(par::mappingParameters)

Reconstruct the 2D grid used for mapping.

SPHtoGrid.get_map_grid_3D — Method

get_map_grid_3D(par::mappingParameters)

Reconstruct the 3D grid used for mapping.

SPHtoGrid.kinetic_SZ — Function

kinetic_SZ(n_cm3::Real, vel_y_cgs::Real, 
                ν::Real=1.e9, z::Real=0.0; 
                DI_over_I::Bool=false)

Computes the kinetic Sunyaev-Zel'dovich effect from electron density n_cm3 and velocity in y-direction to the projection plane in cgs units vel_y_cgs. If DI_over_I is set to true you also need to provide an observation frequency ν and redshift z.

SPHtoGrid.part_weight_XrayBand — Function

part_weight_XrayBand(T::Array{<:Real}, Emin::Real, Emax::Real)

Computes Xray weighted emission of a defined energy band. Emin and Emax are energies in eV.

SPHtoGrid.part_weight_emission — Method

part_weight_emission(rho::Array{<:Real}, T::Array{<:Real})

Emission weighted mapping. Takes density and temperature and computes weights.

SPHtoGrid.part_weight_one — Method

part_weight_one(N::Integer)

Equivalent to no weighting. Returns an Array of ones.

SPHtoGrid.part_weight_physical — Method

part_weight_physical(N::Integer, par::mappingParameters)

Physical weighting function.

SPHtoGrid.project_along_axis — Function

project_along_axis!(x::Array{<:Real}, projection_axis::Integer=3)

Projects and array of 3D along one of the principle axes. projection_axis ∈ {1, 2, 3} => x, y, z axis.

SPHtoGrid.project_along_axis! — Function

project_along_axis!(x::Array{<:Real}, projection_axis::Integer=3)

Projects and array of 3D along one of the principle axes. projection_axis ∈ {1, 2, 3} => x, y, z axis.

SPHtoGrid.project_along_axis! — Function

project_along_axis!(x::Array{<:Real}, projection_axis::Integer=3)

Projects and array of 3D along one of the principle axes. projection_axis ∈ {1, 2, 3} => x, y, z axis.

SPHtoGrid.read_fits_image — Method

read_fits_image(filename::String)

Read a FITS file and return the image, mappingParameters and the snapshot number.

Returns

image: A 2D array with the image pixels

par: mappingParameters used for the image

snap: Number of the mapped snapshot

SPHtoGrid.read_smac1_binary_image — Method

read_smac1_binary_image(filename::String)

Reads a binary image file from Smac and returns a Matrix with the pixel values.

SPHtoGrid.read_smac1_binary_info — Method

read_smac1_binary_info(filename::String)

Returns the image info in a Smac1ImageInfo struct.

SPHtoGrid.rotate_3D! — Method

rotate_3D!(x::Array{<:Real}, alpha::Real, beta::Real, gamma::Real)

Rotates and array of 3D positions around the euler angles α, β and γ corresponding to rotations around the x, y, and z-axis respectively. α, β and γ need to be given in degrees.

SPHtoGrid.rotate_3D — Method

rotate_3D(x::Array{<:Real}, alpha::Real, beta::Real, gamma::Real)

Rotates and array of 3D positions around the euler angles α, β and γ corresponding to rotations around the x, y, and z-axis respectively. α, β and γ need to be given in degrees.

SPHtoGrid.sphMapping — Method

sphMapping( Pos::Array{<:Real}, Bin_Quant::Array{<:Real};
            param::mappingParameters,
            show_progress::Bool=true,
            dimensions::Int=2)

Maps the data in Bin_Quant to a grid using triangualar shaped cloud (TSC) interpolation.

Arguments

Pos: Array with particle positions.

Bin_Quant: Array with particle quantity to be mapped.

show_progress::Bool=true: Show progress bar.

filter_particles::Bool=true: If particles should be filtered before mapping.

dimensions::Int=2: Number of mapping dimensions (2 = to grid, 3 = to cube).

SPHtoGrid.sphMapping — Method

function sphMapping(Pos::Array{T}, HSML::Array{T}, M::Array{T}, 
                    ρ::Array{T}, Bin_Quant::Array{T}, 
                    Weights::Array{T}=ρ;
                    param::mappingParameters,
                    kernel::SPHKernel [,
                    show_progress::Bool=true,
                    parallel::Bool=false,
                    filter_particles::Bool=true,
                    dimensions::Int=2])

Maps the data in Bin_Quant to a grid. Parameters of mapping are supplied in param and the kernel to be used in kernel.

Arguments

Pos: Matrix (3xNpart) with particle positions.

HSML: Array with particle hsml.

M: Array with particle masses.

ρ: Array with particle densities.

Bin_Quant: Array with particle quantity to be mapped.

Weights: Array with weights. Defaults to density-weighted.

kernel::SPHKernel: Kernel object to be used.

show_progress::Bool=true: Show progress bar.

parallel::Bool=true: Run on multiple processors.

filter_particles::Bool=true: Find the particles that are actually contained in the image.

dimensions::Int=2: Number of mapping dimensions (2 = to grid, 3 = to cube).

SPHtoGrid.thermal_SZ — Function

thermal_SZ(n_cm3::Real, T::Real, z::Real=0.0, ν::Real=1.44e9)

Computes the thermal Sunyaev-Zel'dovich effect for electron density n_cm3 and temperature T in Kelvin at redshift z and observer frequency ν.

SPHtoGrid.write_fits_image — Method

write_fits_image(filename::String, image::Array{<:Real}, 
                        par::mappingParameters; 
                        units::String="[i.u.]",
                        snap::Integer=0)

Writes a mapped image to a FITS file and stored the essential mapping parameters in the header.

SPHtoGrid.write_smac1_par — Function

write_smac1_par([...])

Writes a Smac parameter file. Not all relevant parameters implemented yet!

SPHtoGrid.write_smac2_par — Function

write_smac2_par([...])

Writes a P-Smac2 parameter file. Not all relevant parameters implemented yet!

SPHtoGrid.x_ray_emission — Method

x_ray_emission(n_cm3::Real, T_K::Real; 
               Emin::Real=5.e4, Emax::Real=1.e10, 
               xH::Real=0.76)

X-Ray emission of a particle with number density n_cm3 in $1/cm^3$ and temperature T in $K$. Emin and Emax give the minimum and maximum energy of the oberservation. xH gives the hydrogen fraction used in the simulation.

Returns

X-Ray surface brightness contribution in units of [erg/cm^2/s/Hz].

Exported Types

SPHtoGrid.mappingParameters — Type

struct mappingParameters
    x_lim::Vector{Float64}
    y_lim::Vector{Float64}
    z_lim::Vector{Float64}
    center::Vector{Float64}
    pixelSideLength::Float64
    pixelArea::Float64
    Npixels::Vector{Int64}
    x_size::Float64
    y_size::Float64
    z_size::Float64
end

Constructor:

mappingParameters(;x_lim::Vector{Float64}   = [-1.0, -1.0],
                   y_lim::Vector{Float64}   = [-1.0, -1.0],
                   z_lim::Vector{Float64}   = [-1.0, -1.0],
                   center::Vector{Float64}  = [-1.0, -1.0, -1.0],
                   x_size::Float64          =  -1.0,
                   y_size::Float64          =  -1.0,
                   z_size::Float64          =  -1.0,
                   pixelSideLength::Float64 =  -1.0,
                   Npixels::Int64           =   0)

Parameter object for sph to grid mapping. Define either *_lim, or center and *_size. Resolution is defined by pixelSideLength or Npixels.

Private Functions

SPHtoGrid.CS14_acc — Method

CS14_acc(M::Real)

Efficiency from Caprioli&Spitkovsky 2015, doi: 10.1088/0004-637x/783/2/91 Same simplified approach as Vazza+12 -> is roughly half the efficiency of Kang&Ryu 2013.

SPHtoGrid.EpsNtherm — Method

EpsNtherm(rho_cgs::Real, T_K::Real)

Thermal energy density in cgs.

SPHtoGrid.KR07_acc — Method

KR07_acc(M::Float64)

Efficiency model from Kang, Ryu, Cen, Ostriker 2007, http://arxiv.org/abs/0704.1521v1

SPHtoGrid.KR13_acc — Method

KR13_acc(M::Real)

Efficiency model from Kang&Ryu 2013, doi:10.1088/0004-637X/764/1/95

SPHtoGrid.P16_acc — Method

P16_acc(M::Real)

Constant efficiency as in Pfrommer+ 2016, doi: 10.1093/mnras/stw2941

SPHtoGrid.Ryu19_acc — Method

Ryu19_acc(M::Real)

Efficiency model from Ryu et al. 2019, https://arxiv.org/abs/1905.04476 Values for 2.25 < M <= 5.0 extrapolated to entire range

SPHtoGrid.Tcmb — Method

Tcmb(z::Real)

Computes the temperature of the CMB at redshift z.

SPHtoGrid.calculate_index_2D — Method

function calculate_index_2D(i::Integer, j::Integer, x_pixels::Integer)

Calculates the index of a flattened 2D image array.

SPHtoGrid.calculate_index_3D — Method

function calculate_index_3D(i::Integer, j::Integer, x_pixels::Integer)

Calculates the index of a flattened 3D image array.

SPHtoGrid.calculate_weights_2D — Method

function calculate_weights_2D(  wk::Array{<:Real,1}, 
                                iMin::Integer, iMax::Integer, 
                                jMin::Integer, jMax::Integer,
                                x::Real, y::Real, hsml::Real, hsml_inv::Real,
                                kernel::SPHKernel,
                                x_pixels::Integer )

Calculates the kernel- and geometric weights of the pixels a particle contributes to.

SPHtoGrid.calculate_weights_3D — Method

function calculate_weights_3D(  wk::Array{<:Real,1}, 
                                iMin::Integer, iMax::Integer, 
                                jMin::Integer, jMax::Integer,
                                kMin::Integer, kMax::Integer,
                                x::Real, y::Real, z::Real, 
                                hsml::Real, hsml_inv::Real,
                                kernel::SPHKernel,
                                x_pixels::Integer, y_pixels::Integer )

Calculates the kernel- and geometric weights of the pixels a particle contributes to.

SPHtoGrid.check_center_and_move_particles — Method

check_center_and_move_particles(x, par::mappingParameters)

Mapping only works if all coordinates are positive. This function shifts the particles accordingly.

SPHtoGrid.check_in_image — Method

check_in_image(x::Real, y::Real, z::Real, hsml::Real,
                            halfXsize::Real, halfYsize::Real, halfZsize::Real)

Checks if a particle is in the image frame.

SPHtoGrid.comptonY — Method

comptonY(n_cm3::Real, T_K::Real, z::Real)

Computes the Compton-Y parameter from electron density n_cm3 and temperature T in Kelvin at redshift z.

SPHtoGrid.cre_spec_norm_particle — Method

cre_spec_norm_particle(M::Real, eff_function::Function)

Computes the CR electron norm of the particles. This depends on the Mach number M and the acceleration efficiency given by eff_function.

SPHtoGrid.domain_decomposition — Method

domain_decomposition(N::Int64, N_workers::Int64)

Calculate relevant array slices for each worker. Could be done better!

SPHtoGrid.dsa_spectral_index — Method

dsa_spectral_index(M::Real)

Spectral index given by standard Diffuse-Shock-Acceleration (DSA).

SPHtoGrid.euler_matrix — Method

euler_matrix(α::Real, β::Real, γ::Real)

Returns the rotation matrix A based on rotation along the euler angles α, β and γ corresponding to rotations around the x, y, and z-axis respectively.

SPHtoGrid.find_position_periodic — Method

find_position_periodic( pos::Array{<:Real}, k::Integer, boxsize::Real)

Performs a periodic mapping of the particle position.

SPHtoGrid.get_d_hsml_2D — Method

get_d_hsml_2D( dx::Real, dy::Real,
               hsml_inv::Real )

Computes the distance in 2D to the pixel center in units of the kernel support.

SPHtoGrid.get_d_hsml_3D — Method

get_d_hsml_3D( dx::Real, dy::Real, dz::Real,
               hsml_inv::Real )

Computes the distance in 3D to the pixel center in units of the kernel support.

SPHtoGrid.get_dxyz — Method

function get_dxyz(x::Real, hsml::Real, i::Integer)

Calculates the extent of the current particle size in units of pixels.

SPHtoGrid.get_ijk_min_max — Method

function get_ijk_min_max( x::Real, hsml::Real,
                          x_pixels::Integer)

Calculates the minimum and maximum pixel to which a particle contributes.

SPHtoGrid.get_xyz — Method

function get_xyz( pos::Vector{<:Real}, hsml::Real, k::Integer,
                  len2pix::Real, x_pixels::Integer, y_pixels::Integer, z_pixels::Integer,
                  boxsize::Real, periodic::Bool,
                  halfXsize::Real, halfYsize::Real, halfZsize::Real,
                  Ndim::Integer)

Calculates x, y, z position in units of pixels and performs periodic mapping, if required.

SPHtoGrid.kSzPrefac — Method

kSzPrefac(ν::Real, z::Real, DI_over_I::Bool)

Prefactor for the kinetic Sunyaev-Zel'dovich effect.

SPHtoGrid.kr_fitting_function — Method

kr_fitting_function(x::Real, 
                         a0::Real, a1::Real, a2::Real, a3::Real, a4::Real)

Helper function to use the fitting function from KR07.

SPHtoGrid.part_weight_spectroscopic — Method

part_weight_spectroscopic(rho::Array{<:Real}, T::Array{<:Real})

Spectroscopic weighted mapping from Mazotta+ 04. Takes density and temperature and computes weights.

SPHtoGrid.reduce_image_2D — Method

function reduce_image_2D( image::Array{<:Real},
                          x_pixels::Int64, y_pixels::Int64)

Unflattens an image array to a 2D array of pixels.

SPHtoGrid.reduce_image_3D — Method

function reduce_image_3D( image::Array{<:Real}, w_image::Array{<:Real},
                                        x_pixels::Int64, y_pixels::Int64, z_pixels::Int64)

Unflattens an image array to a 3D array of pixels.

SPHtoGrid.rotate_3D_quantity — Method

rotate_3D_quantity(x, α, β, γ)

Rotates a 3D vector along the x-axis with angle α, y-axis with angle β and z-axis with angle γ.

SPHtoGrid.rotate_to_xz_plane! — Method

rotate_to_xz_plane!(x::Array{<:Real})

Rotates an array of 3D positions into the xz-plane.

SPHtoGrid.rotate_to_xz_plane! — Method

rotate_to_xz_plane!(x::Array{<:Real})

Rotates an array of 3D positions into the xz-plane.

SPHtoGrid.rotate_to_yz_plane! — Method

rotate_to_yz_plane(x::Array{<:Real}, x_in::Array{<:Real})

Rotates an array of 3D positions into the yz-plane.

SPHtoGrid.rotate_to_yz_plane! — Method

rotate_to_yz_plane(x::Array{<:Real})

Rotates an array of 3D positions into the yz-plane.

SPHtoGrid.sphMapping_2D — Method

sphMapping2D( Pos::Array{<:Real}, HSML::Array{<:Real}, M::Array{<:Real}, Rho::Array{<:Real}, BinQ::Array{<:Real}, Weights::Array{<:Real}=ones(length(Rho)); param::mappingParameters, kernel::SPHKernel, show_progress::Bool=false )

Underlying function to map SPH data to a 2D grid.

SPHtoGrid.sphMapping_3D — Method

sphMapping2D( Pos::Array{<:Real}, HSML::Array{<:Real}, M::Array{<:Real}, Rho::Array{<:Real}, BinQ::Array{<:Real}, Weights::Array{<:Real}=ones(length(Rho)); param::mappingParameters, kernel::SPHKernel, show_progress::Bool=false )

Underlying function to map SPH data to a 3D grid.

SPHtoGrid.tSzPrefac — Method

tSzPrefac(ν::Real, z::Real)

Computes the prefactor for the thermal Sunyaev-Zel'dovich effect.

SPHtoGrid.update_image — Method

function update_image( image::Real, w_image::Real, 
                       wk::Real, bin_q::Real, 
                       geometry_norm::Real )

Applies the different contributions to the image and the weight image.

Private Types

SPHtoGrid.Smac1ImageInfo — Type

struct Smac1ImageInfo

Stores the information in a Smac binary image header.